+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Improved the building of executable
+
 ## 0.1.0 - 2018-09-21
 ### Added
 - Functionality to download MERRA-2 data using given coordinates
+
+[Unreleased]: https://github.com/Spine-project/MetReLoad/compare/v0.1.0...HEAD